requirements.txt
django==2.2.7
pyjwt==1.7.1
